Created New Panel for our Container. Named: CreateAccountPanel
we are going to use this for the managers activity after they have logged into our program.

currently are having an issue with validating whether a manager or a user have been logged in so for the sake of the demo of the GUI
we have put the call

c1.show(panelCont, "3");

outside of the logic statement that is giving us issues.

I have also placed a TODO in where I am having this issue.