User-Account-Panel
-Small Update

Completely added and tested both deposit and withdraw methods for user account.

Also made sure to surround this with a try and catch block. Rather than the program crashing if a NumberFormatException gets thrown we simply are going to prompt the user to Re-enter an amount if one of our casts gets funny.

TODO:
Continue on hooking up the transfer method.