Added visual aid for the user to see current account standings after everything they do in the inner menu.
Nik i have taken the time to create a clone of the userAcctPanel because I think that my code does not have any of the recent updates to the manager account panel or anything else.

let me know if you want me to send you that copy if there are indeed conflicting with our two versions of the code.

started to hook up transfer but wanted to use the jcombobox that i saw earlier that the manager account had but then i realized i think we had a massive conflict with our master branch.

I will shoot for the complete hookup of all of the panels this weekend so that we can have a few days next week to test for bugs and redundency.